requirements.txt
celery==5.4.0
Flask==3.0.3
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.2.0
pymongo==4.10.1